Bug/420 check duplicate friend request accept

+++ app/Http/Controllers/FriendController.php
@@ -11,6 +11,7 @@ use App\Http\Resources\UserResource;
 use App\Helpers\AchievementHandler;
 use App\Helpers\ActionTrackingHandler;
 use App\Helpers\NotificationHandler;
+use App\Http\Resources\FriendResource;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
@@ -18,8 +19,8 @@ use Illuminate\Http\Response;
 
 class FriendController extends Controller
 {
-    public function show(){
-        // #56
+    public function getFriends(){
+        return new JsonResponse(['friends' => FriendResource::collection(Auth::user()->friends->sortBy('username'))]);
     }
 
     /**
@@ -33,7 +34,7 @@ class FriendController extends Controller
         
         ActionTrackingHandler::handleAction($request, 'DELETE_FRIEND', 'Friendship removed');
         return new JsonResponse(['message' => ['info' => 'Friend removed.'], 
-            'user' => new UserResource(Auth::user())], 
+            'friends' => FriendResource::collection(Auth::user()->friends->sortBy('username'))], 
             Response::HTTP_OK);
     }
 
@@ -62,6 +63,8 @@ class FriendController extends Controller
      * Returns all open requests after updating
      */
     public function acceptFriendRequest(Request $request, Friend $friend){
+        if ($friend->accepted)
+            return new JsonResponse(['message' => ['error' => 'You have already accepted this request.']], Response::HTTP_UNPROCESSABLE_ENTITY);
         $friend->accepted = true;
         $friend->update();
         Friend::create(['user_id' => $friend->friend_id, 'friend_id' => $friend->user_id, 'accepted' => true]);
@@ -72,7 +75,7 @@ class FriendController extends Controller
 
         $requests = $this->fetchRequests();
         return new JsonResponse(['message' => ['success' => 'Friend request accepted. You are now friends.'], 
-            'user' => new UserResource(Auth::user()),
+            'friends' => FriendResource::collection(Auth::user()->friends->sortBy('username')),
             'requests' => $requests], 
             Response::HTTP_OK);
     }
@@ -82,6 +85,8 @@ class FriendController extends Controller
      * Returns the updated list of request
      */
     public function denyFriendRequest(Request $request, Friend $friend){
+        if ($friend->accepted)
+            return new JsonResponse(['message' => ['error' => 'You have already accepted this request.']], Response::HTTP_UNPROCESSABLE_ENTITY);
         $friend->delete();
         $requests = $this->fetchRequests();
         ActionTrackingHandler::handleAction($request, 'FRIEND_REQUEST', 'Friend request denied');
